CORE-3022: Tidy of message patterns refactor
