+++ libs/messaging/messaging-impl/src/main/kotlin/net/corda/messaging/utils/ConfigUtils.kt
@@ -14,7 +14,6 @@ import net.corda.messaging.properties.ConfigProperties.Companion.TOPIC
 import org.osgi.framework.Bundle
 import org.osgi.framework.FrameworkUtil
 import java.net.URL
-import java.util.*
 
 
 class ConfigUtils {
@@ -74,53 +73,10 @@ class ConfigUtils {
     }
 }
 
-/**
- * Read content of a [config] at a given [configPrefix] and its subsections as Java properties.
- * This is usable for generic configuration
- * of Kafka consumer/producers.
- * Override the values of the config with values from [overrideProperties].
- * @param config type safe config
- * @param configPrefix optional key prefix to read config from
- * @param overrideProperties properties to override at the given [configPrefix] in the [config].
- * @return properties with the same content as in config object, with prefix stripped.
- * Keys and values are strings with values overridden by overrideProperties
- */
-fun mergeProperties(
-    config: Config,
-    configPrefix: String?,
-    overrideProperties: Map<String, String>
-): Properties {
-    val properties = Properties()
-    val configAtPrefix = if (configPrefix != null) {
-        config.getConfig(configPrefix)
-    } else {
-        config
-    }
-    configAtPrefix.entrySet().forEach { (key) ->
-        properties.setProperty(
-            key,
-            configAtPrefix.getString(key)
-        )
-    }
-    properties.putAll(overrideProperties)
-    return properties
-}
-
 
 fun Config.getStringOrNull(path: String) = if (hasPath(path)) getString(path) else null
-fun Config.toProperties(): Properties = mergeProperties(this, null, emptyMap())
 fun Config.render(): String =
     root().render(ConfigRenderOptions.defaults().setOriginComments(false).setComments(false).setJson(false))
-
-fun Config.toPatternProperties(pattern: String, clientType: String? = null): String {
-    val pathEnd = if (clientType != null) {
-        "$pattern.$clientType"
-    } else {
-        pattern
-    }
-    return getConfig("messaging.pattern.$pathEnd")
-        .toProperties().entries.sortedBy { it.key.toString() }.joinToString("\n")
-}
 
 fun SubscriptionConfig.toConfig(): Config {
     return ConfigFactory.empty()
