+++ testing/message-patterns/build.gradle
@@ -86,7 +86,7 @@ def kafkaTestingBundle = tasks.register('kafkaTestingBundle', Bundle) {
 
     bnd """\
 Bundle-SymbolicName: \${task.archiveBaseName}-\${task.archiveClassifier}
-    Test-Cases: net.corda.messaging.kafka.integration.publisher.PublisherIntegrationTest
+    Test-Cases: \${classes;HIERARCHY_INDIRECTLY_ANNOTATED;org.junit.platform.commons.annotation.Testable;CONCRETE} 
 """
 }
